+++ CHANGES.md
@@ -4,6 +4,7 @@
 - DMS: Added option `ignore_ddl` to ignore DDL events
 - DMS: Added direct column mapping
 - DMS: Started using direct column mapping by default
+- DMS: Started using the mapping value `OBJECT` when defining column types
 
 ## 2025/06/23 v0.0.23
 - Dependencies: Migrated from `zyp` to `tikray`. It's effectively the